requirements.txt
apscheduler>=3.11.2
cryptography>=43.0.0
dynaconf>=3.2.12
fastapi>=0.128.0
loguru>=0.7.3
minio>=7.2.0
modelscope>=1.33.0
numpy>=2.1.0,<2.4.0
openai>=2.14.0
pandas>=2.3.3
pgvector>=0.4.2
pillow>=12.0.0
plotly>=6.5.0
psutil>=6.1.0
psycopg[binary]>=3.3.2
pymysql>=1.1.2
pytest>=9.0.2
pytest-cov>=7.0.0
pytest-mock>=3.15.1
python-dateutil>=2.9.0.post0
pytz>=2024.2
pyyaml>=6.0.2
redis>=7.1.0
requests>=2.32.0
rich>=14.0.0
scikit-learn>=1.8.0
sqlalchemy>=2.0.45
sqlalchemy-utils>=0.42.1
streamlit>=1.52.2
tabulate>=0.9.0
tenacity>=9.1.2
torch>=2.9.0
transformers>=4.57.3
uvicorn>=0.40.0
